Splitting Types.hpp into separate Matrix, Vector and Color headers

+++ Source/Rasterizer/Attribute.hpp
@@ -6,10 +6,10 @@ namespace Rasterizer
 	struct TAttribute
 	{
 		const void* Pointer = nullptr;
-		uint32_t Stride = sizeof(T);
+		size_t Stride = sizeof(T);
 
 		// Get attribute by vertex index
-		const T& operator[](uint32_t i) const
+		const T& operator[](size_t i) const
 		{
 			return *(const T*)((const char*)(Pointer) + Stride * i);
 		}
