+++ Source/Rasterizer/Mesh.hpp
@@ -1,7 +1,8 @@
 #pragma once
 
-#include "Types.hpp"
 #include "Attribute.hpp"
+#include "Util/Vector.hpp"
+#include <cstdint>
 
 namespace Rasterizer
 {
@@ -10,6 +11,6 @@ namespace Rasterizer
 		TAttribute<FVector3f> Positions = {};
 		TAttribute<FVector4f> Colors = {};
 		uint32_t const* Indices = nullptr;
-		uint32_t Count = 0; // If Indices is valid, contains the number indices in the array. If Indices is not valid, count is the number of vertices.
+		size_t Count = 0; // If Indices is valid, contains the number indices in the array. If Indices is not valid, count is the number of vertices.
 	};
 }
